refactor : refactor component directory structure
