+++ src/components/atoms/Icon/Icon.test.tsx
